+++ src/components/render-clip/render-clip.js
@@ -21,6 +21,10 @@ export default class RenderClip extends React.Component {
       }
     };
 
+    this.onWheel = () => {
+      this.props.controller.onWheel();
+    };
+
     this.syncHeightAfterRender = () => {
       const {renderedItemsCount, orthogonalScrollSizeProp, onRender} = this.props.controller;
 
@@ -122,7 +126,8 @@ export default class RenderClip extends React.Component {
           overflow: 'auto',
           whiteSpace: 'nowrap'
         },
-        onScroll: this.onScroll
+        onScroll: this.onScroll,
+        onWheel: this.onWheel
       },
       !!renderedItemsCount && createElement(
         'div',