+++ src/controllers/render-clip/render-clip-1d-definition.js
@@ -25,6 +25,7 @@ export default ({
       'setDomContainer',
       'setItems',
       'setItemsCount',
+      'onWheel',
       'onScroll',
       'scrollIntoView',
       'scrollToFit',